CommonLib, ClientLib, Data: 부동 소수점 오차를 고려해 위치, 거리, 이동 시스템을 정수 기반으로 변경

+++ Assets/Script/CommonLib/EntityData.cs
@@ -14,7 +14,7 @@ namespace Script.CommonLib
         public uint maxHp;
         public uint attackDamage;
         public ushort attackDelayMs;
-        public byte attackRange;
+        public ushort attackRange;
         public ushort moveSpeed;
     }
 }
